QDate implementation & guestDialog proper implementation

+++ newguestdialog.cpp
@@ -18,7 +18,8 @@ newGuestDialog::~newGuestDialog()
 
 void newGuestDialog::setnewGuestData()
 {
+
     newGuestName=ui->lineEdit->text();
-    roomNumSet=ui->lineEdit_2->text();
-    newNumNights=ui->lineEdit_3->text();
+    roomNumSet=ui->lineEdit_2->text().toInt();
+    newNumNights=ui->lineEdit_3->text().toInt();
 }
